# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SPARC API
  version: 1.1.0-alpha.1
  summary: Contract-first API for satellite-derived local environmental proxy indicators.
  description: |
    The read-only P0 demo API is implemented over immutable contract examples; live raster
    processing and administrative job creation remain disabled.
    SPARC results are satellite-derived proxy indicators, not official UN SDG indicators,
    and do not establish causation. `loading` is a browser state; server resources use
    queued, running, complete, partial, failed, or cancelled states.

    Validation duties that remain authoritative server responsibilities:

    1. Identifier patterns constrain shape only. IDs are resolved through approved catalogues,
       never concatenated into filesystem paths or outbound URLs.
    2. Calendar dates are parsed as real dates and checked for ordering, supported archive
       bounds, maximum span, and period compatibility before any I/O.
    3. Public comparison requests resolve bounded precomputed results only. They cannot queue
       anonymous live processing and remain subject to request-size, concurrency, and rate limits.
    4. Problem responses never include stack traces, filesystem paths, environment values,
       internal hostnames, credentials, tokens, signed URLs, or raw provider responses.

    Forecast endpoints are read-only and currently serve a versioned prototype
    pack. Their probabilities are transparent baseline estimates and are not
...
servers:
...
security: []
paths:
...
  /api/v1/regions:
    get:
      operationId: listRegions
      tags: [Regions]
      summary: List available districts and child regions with capabilities.
      x-sparc-maturity: required-p0
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [district, subdistrict, block]
        - name: parentId
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*(?::[a-z0-9]+(?:-[a-z0-9]+)*)*$'
      responses:
        '200':
          description: Region references; large geometries are linked rather than embedded.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta, links]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/RegionRef'
                  meta:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ResponseMeta'
                  links:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/Links'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}:
    get:
      operationId: getRegion
      tags: [Regions]
      summary: Return one region, its parent/children links, bounds, and capabilities.
      x-sparc-maturity: required-p0
      parameters:
        - $ref: '#/components/parameters/RegionId'
      responses:
        '200':
          description: Region metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta, links]
                properties:
                  data:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/RegionRef'
                  meta:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ResponseMeta'
                  links:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/Links'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/forecast-runs:
    get:
      operationId: listForecastRuns
      tags: [Forecasts]
      summary: List published forecast runs for a region.
      description: Returns immutable run metadata for the flood, drought, and heat prototype products.
      x-sparc-maturity: implemented-prototype
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - name: hazard
          in: query
          required: false
          schema:
            $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ForecastHazard'
      responses:
        '200':
          description: Published forecast-run metadata.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ForecastRunListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/forecasts/{hazard}/latest:
    get:
      operationId: getLatestForecast
      tags: [Forecasts]
      summary: Return the latest published forecast for one hazard.
      description: The response contains horizon points, separate probability and confidence fields, drivers, provenance, and limitations.
      x-sparc-maturity: implemented-prototype
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/ForecastHazard'
      responses:
        '200':
          description: Latest forecast run.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ForecastRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/forecasts/{hazard}/{runId}:
    get:
      operationId: getForecastRun
      tags: [Forecasts]
      summary: Return one immutable forecast run.
      x-sparc-maturity: implemented-prototype
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/ForecastHazard'
        - $ref: '#/components/parameters/ForecastRunId'
      responses:
        '200':
          description: Forecast run with all supported horizons.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ForecastRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/forecasts/{hazard}/{runId}/timeseries:
    get:
      operationId: getForecastTimeSeries
      tags: [Forecasts]
      summary: Return the horizon series for one immutable forecast run.
      x-sparc-maturity: implemented-prototype
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/ForecastHazard'
        - $ref: '#/components/parameters/ForecastRunId'
      responses:
        '200':
          description: Forecast probability and metric time series.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ForecastTimeSeriesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/summary:
    get:
      operationId: getRegionSummary
      tags: [Regions, Indicators]
      summary: Return a district or child-region summary for two supported periods.
      x-sparc-maturity: required-p0
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/BaselineStart'
        - $ref: '#/components/parameters/BaselineEnd'
        - $ref: '#/components/parameters/ComparisonStart'
        - $ref: '#/components/parameters/ComparisonEnd'
      responses:
        '200':
          description: Complete or usable partial summary. Inspect `meta.partial` and warnings.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/DistrictSummaryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/regions/{regionId}/indicators:
    get:
      operationId: listRegionIndicators
      tags: [Indicators]
      summary: List supported indicator results for a region and period pair.
      x-sparc-maturity: required-p0
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/BaselineStart'
        - $ref: '#/components/parameters/BaselineEnd'
        - $ref: '#/components/parameters/ComparisonStart'
        - $ref: '#/components/parameters/ComparisonEnd'
        - name: indicatorId
          in: query
          required: false
          schema:
            $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/IndicatorId'
      responses:
        '200':
          description: Summary list; an unavailable indicator is explicit rather than silently omitted.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta, links]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/IndicatorSummary'
                  meta:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/ResponseMeta'
                  links:
                    $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/Links'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/regions/{regionId}/indicators/{indicatorId}:
    get:
      operationId: getRegionIndicator
      tags: [Indicators]
      summary: Return one detailed indicator comparison, quality, provenance, interpretation, and layers.
      x-sparc-maturity: required-p0
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/IndicatorId'
        - $ref: '#/components/parameters/BaselineStart'
        - $ref: '#/components/parameters/BaselineEnd'
        - $ref: '#/components/parameters/ComparisonStart'
        - $ref: '#/components/parameters/ComparisonEnd'
      responses:
        '200':
          description: Detailed complete or partial result.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/IndicatorComparisonResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/regions/{regionId}/timeseries:
    get:
      operationId: getRegionTimeSeries
      tags: [Indicators]
      summary: Return a same-method time series for one indicator.
      x-sparc-maturity: planned-p1
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - name: indicatorId
          in: query
          required: true
          schema:
            $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/IndicatorId'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
            maxLength: 10
            pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
            maxLength: 10
            pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
      responses:
        '200':
          description: Time series with provenance.
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/TimeSeriesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/comparisons/{comparisonId}:
    get:
      operationId: getComparison
      tags: [Comparisons]
      summary: Return the current comparison summary.
      x-sparc-maturity: required-p0
      parameters:
        - $ref: '#/components/parameters/ComparisonId'
      responses:
        '200':
          description: Complete or usable partial comparison summary.
          content:
            application/json:
              schema:
                $ref: '../packages/contracts/schemas/sparc.schema.json#/$defs/DistrictSummaryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-admin-token
      description: Planned P1 server-side administrative credential; never stored in browser code.
    ReportAccess:
      type: apiKey
      in: header
      name: X-Report-Access
      description: Short-lived opaque P0 report capability token; never a government credential and never placed in a URL.
  headers:
    ETag:
      description: Strong or weak validator for immutable result content.
      schema:
        type: string
...